% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binchain.R
\docType{data}
\name{binchain}
\alias{binchain}
\title{Example VLMC data}
\format{
An object of class \code{list} of length 4.
}
\usage{
binchain
}
\description{
List of 4 binary chains with length 2500 each.
}
\author{
Victor Freguglia
}
\keyword{datasets}
