% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinsic_bf.R
\name{intrinsic_bf_cmp}
\alias{intrinsic_bf_cmp}
\title{Intrinsic Bayes Factor for Renewal State Testing}
\usage{
intrinsic_bf_cmp(
  z,
  renewal,
  nsamples = 20000,
  burnin = 10000,
  Hmax = 5,
  allowedMatrix = NULL,
  alpha0 = 1/2,
  alpha1 = 1/2,
  logpenalty0 = 0,
  logpenalty1 = 0,
  seed = NULL,
  subset_size = 2,
  max_subsets_evaluated = NULL
)
}
\arguments{
\item{z}{A \code{list} object containing independent VLMC sequences in each element.}

\item{renewal}{Which states are considered renewal states under the H0 hypothesis.}

\item{nsamples}{Size of context-tree chains in each internal Metropolis-Hastings algorithm run.}

\item{burnin}{Number of Metropolis-Hastings algorithm steps executed before starting to record states.}

\item{Hmax}{Maximum height of the context tree considered.}

\item{allowedMatrix}{A logical \code{matrix} specifying which transitions are allowed (i.e., have positive probability).
The row \code{i} and column \code{j} indicates whether a transition from (i-1) to (j-1) is allowed.
\code{NULL} can be used to indicate that every transition is allowed.}

\item{alpha0, alpha1}{The Dirichlet parameter to be considered in the prior distribution of probabilities, under
the H0 and H1 hypothesis respectively.}

\item{logpenalty0, logpenalty1}{Penalty considered in the context tree prior distribution for growing a new branch for the current tree. \code{0} represents a uniform context tree prior distribution.}

\item{seed}{Random seed to be used as a reference for the Partial Bayes Factors computations. Passing this argument instead of using the \code{set.seed()} function is required to ensure reproducibility because of parallel computations carried in the function.}

\item{subset_size}{Number of samples used in each training set. Increasing this value may drastically increases the number of combinations of sequences used as training samples, it is not recommended to use values larger than \code{2}.}

\item{max_subsets_evaluated}{Maximum number of Partial Bayes Factors to compute when the number of combinations of the chosen subset size is too large. The selected subsets are selected at random.}
}
\value{
A \code{list} object containing:
\itemize{
\item \code{ibf_arithmetic}: Arithmetic Intrinsic Bayes Factor computed.
\item \code{ibf_geometric}: Geometric Intrinsic Bayes Factor computed.
\item \code{partials}: A list with the output of each Partial Bayes Factor computed.
\item \code{total_time}: Total time elapsed.
}
}
\description{
Intrinsic Bayes Factor for Renewal State Testing
}
\details{
In order to fully take advantage of parallel computing in this function, the user should set a \code{plan} from the \code{future} package.

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(future)
plan(multisession)
}\if{html}{\out{</div>}}
}
\author{
Victor Freguglia
}
